Group parse-args' loop output params in a map
Every param needs to be provided to `recur`, which means that all
`recur` calls need to be updated every time you add or change a param.
Besides that, using functions like `assoc` or `update` makes it easier
to understand which params have been modified and which params remain
the same.